refactor(extract): move utl into extract
as it's currently only used from there
