+++ src/extract/resolve/resolve-AMD.js
@@ -4,7 +4,7 @@ const fs                       = require('fs');
 const path                     = require('path');
 const isCore                   = require('resolve').isCore;
 const memoize                  = require('lodash/memoize');
-const pathToPosix              = require('../../utl/pathToPosix');
+const pathToPosix              = require('../utl/pathToPosix');
 const determineDependencyTypes = require('./determineDependencyTypes');
 const readPackageDeps          = require('./readPackageDeps');
 const resolveHelpers           = require('./resolve-helpers');
