+++ src/extract/resolve/resolve-commonJS.js
@@ -2,7 +2,7 @@
 
 const path                     = require('path');
 const isCore                   = require('resolve').isCore;
-const pathToPosix              = require('../../utl/pathToPosix');
+const pathToPosix              = require('../utl/pathToPosix');
 const determineDependencyTypes = require('./determineDependencyTypes');
 const readPackageDeps          = require('./readPackageDeps');
 const resolveHelpers           = require('./resolve-helpers');
